move function restore into utils::restore
