+++ src-tauri/src/utils/mod.rs
@@ -1,2 +1,5 @@
 pub mod fs;
+mod restore;
 pub mod scan;
+
+pub use restore::restore;
